WIP on skill casts.  Read Notes
Make sure you have a skill unlocked before you try to cast it.  just because you have it on your short-cut bar, doesn't mean the skill is unlocked.

Using some string parsing to chop the middle of the SkillCastID out, and translate it to an EO ID.  works for several spells, but not everything..