Improved monster support
Moved monster move functions to MonsterSpawn.
Better multi client, if monster is already spawned, it will spawn it and send the move for the next coord in pathing.
Changed to player casting, cast animation starts, but need to figure out how to stop it. It won't cast a second time currently and effect stays.
Added a second monster to RDF.
Fix problem with database lookups for monsters and monster coords.